Support rename from readme => readme_renderer
